rename gradientbutton to gradientborder
